+++ apps/dotlottie-web-example/src/main.ts
@@ -5,6 +5,7 @@
 /* eslint-disable no-secrets/no-secrets */
 
 import './styles.css';
+import type { Mode } from '@lottiefiles/dotlottie-web';
 import { DotLottie } from '@lottiefiles/dotlottie-web';
 
 import wasmUrl from '../../../packages/web/dist/renderer.wasm?url';
@@ -13,22 +14,7 @@ const app = document.getElementById('app') as HTMLDivElement;
 
 app.innerHTML = `
 <div class="grid">
-  <canvas data-src="https://lottie.host/1cf72a35-6d88-4d9a-9961-f1bb88087f2c/miJIHiyH4Q.lottie"></canvas>
-  <canvas data-src="https://lottie.host/647eb023-6040-4b60-a275-e2546994dd7f/zDCfp5lhLe.json"></canvas>
-  <canvas data-src="https://lottie.host/a7421582-4733-49e5-9f77-e8d4cd792239/WZQjpo4uZR.lottie"></canvas>
-  <canvas data-src="https://lottie.host/e2a24b6f-df7f-4fc5-94ea-30f0846f85dc/1RLOR2g0m3.lottie"></canvas>
-  <canvas data-src="https://lottie.host/35326116-a8ca-4219-81ca-df9ce56a3f22/zCGFevEA23.lottie"></canvas>
-  <canvas data-src="https://lottie.host/f315768c-a29b-41fd-b5a8-a1c1dfb36cd2/CRiiNg8fqQ.lottie"></canvas>
-  <canvas data-src="/down.json"></canvas>
-  <canvas data-src="/dragon.json"></canvas>
-  <canvas data-src="/editor.json"></canvas>
-  <canvas data-src="/growup.json"></canvas>
-  <canvas data-src="/hamster.lottie"></canvas>
-  <canvas data-src="/like.json"></canvas>
-  <canvas data-src="/lolo.json"></canvas>
-  <canvas data-src="/walker.json"></canvas>
-  <canvas data-src="/waves.json"></canvas>
-  <canvas data-src="/woman.json"></canvas>
+  
 </div>
 
 <div class="container">
@@ -38,14 +24,22 @@ app.innerHTML = `
     <button id="stop" class="control-button">Stop</button>
     
     <label for="frameSlider">Frame: <span id="current-frame">0</span></label>
-    <input type="range" id="frameSlider" min="0" step="1" />
+    <input type="range" id="frameSlider" min="0" step="0.1" />
 
     <label for="loopToggle">Loop: </label>
     <input type="checkbox" id="loopToggle" checked />
     <label for="speed" class="speed-label">Speed: <span id="speed-value">x1</span></label>
     <input type="range" id="speed" min="0.1" max="5" step="0.1" value="1" class="speed-slider" />
+    <button id="jump" class="control-button">Jump</button>
     <button id="destroy" class="control-button" style="background: #cd3434;">Destroy</button>
     <button id="reload" class="control-button">Reload</button>
+    <label for="mode">Mode: </label>
+    <select id="mode">
+      <option value="bounce">Bounce</option>
+      <option value="bounce-reverse">Bounce Reverse</option>
+      <option value="reverse">Reverse</option>
+      <option value="forward">Forward</option>
+    </select>
   </div>
 </div>
 `;
@@ -62,6 +56,7 @@ const allCanvas = document.querySelectorAll('canvas[data-src]') as NodeListOf<HT
 
 allCanvas.forEach((canvas) => {
   const src = canvas.getAttribute('data-src') as string;
+  const backgroundColor = canvas.getAttribute('data-bg-color') as string;
 
   // eslint-disable-next-line no-new
   new DotLottie({
@@ -69,6 +64,7 @@ allCanvas.forEach((canvas) => {
     src,
     loop: true,
     autoplay: true,
+    backgroundColor,
   });
 });
 
@@ -85,8 +81,9 @@ fetch('/hamster.lottie')
       loop: true,
       autoplay: true,
       mode: 'bounce',
-      segments: [10, 14],
+      segments: [10, 90],
       speed: 1,
+      backgroundColor: 'purple',
     });
 
     const playPauseButton = document.getElementById('playPause') as HTMLButtonElement;
@@ -98,6 +95,18 @@ fetch('/hamster.lottie')
     const speedValueSpan = document.getElementById('speed-value') as HTMLSpanElement;
     const destroyButton = document.getElementById('destroy') as HTMLButtonElement;
     const reloadButton = document.getElementById('reload') as HTMLButtonElement;
+    const jumpButton = document.getElementById('jump') as HTMLButtonElement;
+    const modeSelect = document.getElementById('mode') as HTMLSelectElement;
+
+    modeSelect.addEventListener('change', () => {
+      dotLottie.setMode(modeSelect.value.toString() as Mode);
+    });
+
+    jumpButton.addEventListener('click', () => {
+      const midFrame = dotLottie.totalFrames / 2;
+
+      dotLottie.setFrame(midFrame);
+    });
 
     destroyButton.addEventListener('click', () => {
       canvas.remove();
@@ -115,7 +124,7 @@ fetch('/hamster.lottie')
     });
 
     playPauseButton.addEventListener('click', () => {
-      if (dotLottie.playing) {
+      if (dotLottie.isPlaying) {
         dotLottie.pause();
       } else {
         dotLottie.play();
@@ -161,7 +170,7 @@ fetch('/hamster.lottie')
 
     dotLottie.addEventListener('play', (event) => {
       console.log(event);
-      frameSlider.max = (dotLottie.totalFrames - 1).toString();
+      frameSlider.max = dotLottie.totalFrames.toString();
       playPauseButton.innerText = 'Pause';
     });
 
@@ -171,10 +180,10 @@ fetch('/hamster.lottie')
     });
 
     dotLottie.addEventListener('frame', (event) => {
-      const roundedFrame = Math.round(event.currentFrame);
+      const frame = parseFloat(event.currentFrame.toFixed(2));
 
-      frameSlider.value = roundedFrame.toString();
-      currentFrameSpan.textContent = roundedFrame.toString();
+      frameSlider.value = frame.toString();
+      currentFrameSpan.textContent = frame.toString();
     });
 
     dotLottie.addEventListener('loop', (event) => {
@@ -190,11 +199,8 @@ fetch('/hamster.lottie')
 
     dotLottie.addEventListener('stop', (event) => {
       console.log(event);
-      const roundedFrame = Math.round(dotLottie.currentFrame);
 
       playPauseButton.innerText = 'Play';
-      frameSlider.value = roundedFrame.toString();
-      currentFrameSpan.textContent = roundedFrame.toString();
     });
   })
   .catch((error) => {